re-organized low level tdvp operations
