+++ src/mqt/yaqs/digital/digital_tjm.py
@@ -34,7 +34,7 @@ from ..core.methods.decompositions import merge_two_site, split_two_site
 from ..core.methods.dissipation import apply_dissipation
 from ..core.methods.stochastic_process import stochastic_process
 from ..core.methods.tdvp import tdvp
-from ..core.methods.tdvp_retained_bonds import prepare_retained_bonds
+from ..core.methods.tdvp_bond_support import prepare_support_bonds
 from ..core.random_utils import make_trajectory_rng
 from .utils.dag_utils import convert_dag_to_tensor_algorithm
 
@@ -248,12 +248,12 @@ def apply_two_qubit_gate_tdvp(
     short_state, short_mpo, window = apply_window(state, mpo, first_site, last_site, window_size)
 
     site0, site1 = gate.sites[0], gate.sites[1]
-    retained_bonds = (
-        prepare_retained_bonds(short_state, site0, site1, (window[0], window[1]), sim_params)
+    support_bonds = (
+        prepare_support_bonds(short_state, site0, site1, (window[0], window[1]), sim_params)
         if abs(site0 - site1) != 1
         else None
     )
-    tdvp(short_state, short_mpo, sim_params, mode="dynamic", retained_bonds=retained_bonds)
+    tdvp(short_state, short_mpo, sim_params, mode="dynamic", support_bonds=support_bonds)
     for i in range(window[0], window[1] + 1):
         state.tensors[i] = short_state.tensors[i - window[0]]
